飞机大战

+++ src/main/java/com/zhenai/ecsgame/framwork/gameEngine/bean/Position.java
@@ -7,10 +7,10 @@ package com.zhenai.ecsgame.framwork.gameEngine.bean;
  */
 public class Position {
 
-    private double x;
-    private double y;
+    private int x;
+    private int y;
 
-    public Position(double x, double y) {
+    public Position(int x, int y) {
         this.x = x;
         this.y = y;
     }
@@ -20,19 +20,19 @@ public class Position {
         this.y = 0;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
